# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: 연금술사 Prototype API
  version: 0.1.0
paths:
  /api/v1/auth/login:
    post:
      summary: Login
      operationId: login_api_v1_auth_login_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Login'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/auth/refresh:
    post:
      summary: Refresh
      operationId: refresh_api_v1_auth_refresh_post
      parameters:
      - name: refresh
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Refresh
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/auth/logout:
    post:
      summary: Logout
      operationId: logout_api_v1_auth_logout_post
      security:
      - HTTPBearer: []
      parameters:
      - name: refresh
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Refresh
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/me:
    get:
      summary: Me
      operationId: me_api_v1_me_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
      security:
      - HTTPBearer: []
  /api/v1/organizations:
    get:
      summary: Organizations
      operationId: organizations_api_v1_organizations_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
      security:
      - HTTPBearer: []
  /api/v1/projects:
    get:
      summary: Projects
      operationId: projects_api_v1_projects_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
      security:
      - HTTPBearer: []
    post:
      summary: Create Project
      operationId: create_project_api_v1_projects_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectInput'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
  /api/v1/projects/{project_id}:
    get:
      summary: Project Detail
      operationId: project_detail_api_v1_projects__project_id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Project Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    patch:
      summary: Update Project
      operationId: update_project_api_v1_projects__project_id__patch
      security:
      - HTTPBearer: []
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Project Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectInput'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/projects/{project_id}/members:
    get:
      summary: Members
      operationId: members_api_v1_projects__project_id__members_get
      security:
      - HTTPBearer: []
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Project Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      summary: Add Member
      operationId: add_member_api_v1_projects__project_id__members_post
      security:
      - HTTPBearer: []
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Project Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MemberInput'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/projects/{project_id}/members/{user_id}:
    delete:
      summary: Remove Member
      operationId: remove_member_api_v1_projects__project_id__members__user_id__delete
      security:
      - HTTPBearer: []
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Project Id
      - name: user_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: User Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/admin/users/{user_id}:
    patch:
      summary: Change Role
      operationId: change_role_api_v1_admin_users__user_id__patch
      security:
      - HTTPBearer: []
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: User Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoleInput'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/admin/audit:
    get:
      summary: Audit Events
      operationId: audit_events_api_v1_admin_audit_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
      security:
      - HTTPBearer: []
  /api/v1/templates:
    get:
      summary: Templates
      operationId: templates_api_v1_templates_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
      security:
      - HTTPBearer: []
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    Login:
      properties:
        username:
          type: string
          maxLength: 80
          minLength: 1
          title: Username
        password:
          type: string
          maxLength: 200
          minLength: 1
          title: Password
      additionalProperties: false
      type: object
      required:
      - username
      - password
      title: Login
    MemberInput:
      properties:
        user_id:
          type: string
          format: uuid
          title: User Id
        role:
          $ref: '#/components/schemas/Role'
          default: User
      additionalProperties: false
      type: object
      required:
      - user_id
      title: MemberInput
...
    ProjectInput:
      properties:
        name:
          type: string
          maxLength: 120
          minLength: 1
          title: Name
        description:
          type: string
          maxLength: 2000
          title: Description
          default: ''
        security_class:
          type: string
          maxLength: 30
          title: Security Class
          default: synthetic
        template_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Template Id
      additionalProperties: false
      type: object
      required:
      - name
      title: ProjectInput
...
    Role:
      type: string
      enum:
      - User
      - Reviewer
      - OrgAdmin
      - SystemAdmin
      title: Role
    RoleInput:
      properties:
        role:
          $ref: '#/components/schemas/Role'
        active:
          type: boolean
          title: Active
          default: true
      additionalProperties: false
      type: object
      required:
      - role
      title: RoleInput
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer